# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HMS SaaS API Specification",
    "description": "Production-ready REST endpoints for Hospital Management SaaS",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Local API gateway"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authenticate User",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["usernameOrEmail", "password"],
                "properties": {
                  "usernameOrEmail": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login success" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register New User",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password", "firstName", "lastName"],
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "phone": { "type": "string" },
                  "roles": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Registration successful" }
        }
      }
    },
    "/patients": {
      "get": {
        "summary": "List Patients",
        "responses": {
          "200": { "description": "List fetched" }
        }
      }
    },
    "/doctors": {
      "get": {
        "summary": "List Doctors",
        "responses": {
          "200": { "description": "List fetched" }
        }
      }
    },
    "/appointments": {
      "post": {
        "summary": "Book Appointment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["patientId", "doctorId", "appointmentDateTime"],
                "properties": {
                  "patientId": { "type": "integer" },
                  "doctorId": { "type": "integer" },
                  "appointmentDateTime": { "type": "string", "format": "date-time" },
                  "reason": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Appointment booked" }
        }
      }
    }
  }
}